init file driver

+++ myos/allocator.cpp
@@ -7,7 +7,7 @@ VirtPageAllocator* virt_page_allocator;
 alignas(PhysPageAllocator) static uint8_t phy_buf[sizeof(PhysPageAllocator)];
 alignas(VirtPageAllocator) static uint8_t virt_buf[sizeof(VirtPageAllocator)];
 
-void init_allocators(uint8_t* bitmap, uint64_t total_pages) {
+void init_allocators(uint64_t* bitmap, uint64_t total_pages) {
     phy_page_allocator = new (phy_buf) PhysPageAllocator();
     phy_page_allocator->init(bitmap, total_pages);
